backend/requirements.txt
psycopg2
django==4.1.1
djangorestframework~=3.12.4
django-dotenv
pytz
